Added finished Fight Component
